% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_redres.R
\name{compute_redres}
\alias{compute_redres}
\title{Compute residuals for a linear mixed model}
\usage{
compute_redres(model, type = "raw_cond")
}
\arguments{
\item{model}{Model fit using \code{lmer} from \code{lme4}.}

\item{type}{String identifying type of residual. Default is "raw_cond". See details for the options available.}
}
\value{
Returns a vector of residuals according to type specified. Residuals appear in the same order as the observations used to fit the model.
}
\description{
Computes residuals for a given linear mixed model.
}
\details{
Residual types available:
\itemize{
  \item \code{"pearson_cond"}: Pearson conditional residuals
  \item \code{"pearson_mar"}: Pearson marginal residuals
  \item \code{"raw_cond"}: raw conditional residuals (default)
  \item \code{"raw_mar"}: raw marginal residuals
  \item \code{"std_cond"}: studentized conditional residuals
  \item \code{"std_mar"}: studentized marginal residuals
}
See the \href{https://goodekat.github.io/redres/articles/redres-vignette.html#extracting-residuals}{vignette} for details on how residual types are computed.
}
\examples{
# fits a linear mixed effects model
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), data = sleepstudy)

# computes the default residuals (raw conditional)
compute_redres(fm1)

# changes the residual type to studentized marginal
compute_redres(fm1, type = "std_mar")
}
