% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condres.R
\name{condres}
\alias{condres}
\title{condress}
\usage{
condres(model)
}
\arguments{
\item{model}{input mixed model}
}
\description{
This is Function that is used to identify conditional residule of the model.
}
\examples{

# read data from stat510 of Iowa State University
d = read.delim("http://dnett.github.io/S510/SeedlingDryWeight2.txt")
d$Genotype = factor(d$Genotype)
# fit the mixed model
model <- lmer(SeedlingWeight ~ Genotype + (1|Tray), REML = F,data = d)
# using the condres function to get coditional residule of the model
condres(model)
}
