% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_paprika.R
\docType{data}
\name{paprika}
\alias{paprika}
\title{Paprika Dataset}
\format{Tibble}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/25747}
}
\usage{
paprika
}
\description{
Data from a split-plot design study on paprika
}
\keyword{datasets}
