% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genres.R
\name{plot_genres}
\alias{plot_genres}
\title{Generalized residual quantile plot}
\usage{
plot_genres(model)
}
\arguments{
\item{model}{Model fit using \code{lmer}.}
}
\value{
A generalized residual quantile plot.
}
\description{
Use to compare the generalized residuals to the continuous uniform quantiles.
Obvious departures indicate that the assumed model is inadequate to describe the observed data.
}
\details{
Generalized residuals should be approximately Uniform(0,1) distributed if the model is
adequate for the data. See the
\href{https://goodekat.github.io/redres/articles/redres-vignette.html#genearlized}{vignette}
for details on how generalized residuals are computed.
}
\examples{
# fits a linear mixed effect model using lme4 where model has a
# random intercept for Days and random slope for Subject*Days
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)

# check that all model assumptions are reasonable using generalized residuals
plot_genres(fm1)
}
