% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranef.R
\name{plot_ranef}
\alias{plot_ranef}
\title{Normal quantile plots for random effects}
\usage{
plot_ranef(model)
}
\arguments{
\item{model}{Model fit using \code{lmer} from \code{lme4}.}
}
\value{
A grid of normal quantile plots for the random effects.
}
\description{
Plots each random effect in the model against the normal quantiles. For linear mixed models,
random effects are assumed to be normally distributed. This plot can be used to assess the
validity of that assumption.
}
\examples{
# fits a linear mixed effects model
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)

# plots normal quantiles of random effects vector to assess normality
plot_ranef(fm1)
}
