% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raneff.R
\name{plot_raneff}
\alias{plot_raneff}
\title{Normal quantile plots for random effects}
\usage{
plot_raneff(model)
}
\arguments{
\item{model}{Model fit using \code{lmer}.}
}
\value{
A normal quantile plot for the random effects.
}
\description{
Plots the vector of all random effects against the normal quantiles. For linear mixed models
random effects are assumed to be normally distributed. This plot can be used to assess the
validity of that assumption.
}
\examples{
# fits a linear mixed effects model
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)

# plots normal quantiles of random effects vector to assess normality
plot_raneff(fm1)
}
