% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resid.R
\name{plot_redres}
\alias{plot_redres}
\title{Diagnostic Residual Plot}
\usage{
plot_redres(model, type = "raw_cond")
}
\arguments{
\item{model}{Model fit using \code{lmer}.}

\item{type}{String identifying type of residual. Default is "raw_cond". See \code{\link{redres}} for details of available types.}
}
\value{
A residual plot.
}
\description{
Plots specified residual type by fitted values for a linear mixed effects model
fitted using \code{lmer}. Plot used to assess whether the assumptions of constant
variance and linear form are adequate.
}
\examples{
# Fit a linear mixed effect model with a default (raw conditional) residual type.
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
# Plot raw conditional residuals by fitted values.
plot_redres(fm1)
# Plot standardized conditional residuals by fitted values.
plot_redres(fm1, type = "std_cond")
}
