% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_redres.R
\name{redres}
\alias{redres}
\title{Compute residuals for a linear mixed model}
\usage{
redres(model, type = "raw_cond")
}
\arguments{
\item{model}{Model fit using \code{lmer} from lme4 for which residuals will be computed.}

\item{type}{String identifying type of residual. Default is "raw_cond". See details for the options available.}
}
\value{
Returns residuals according to type specified.
}
\description{
Computes residuals for linear mixed models fit using \code{lmer} from the lme4.
It can return seven types of residuals inlcuding options that are not available
from lme4. These types are conditional raw, Pearson, and studentized, marginal
raw, Pearson, and studentized, and generalized residuals.
}
\details{
Residual types available
\itemize{
  \item \code{"genres"}: generalized residuals
  \item \code{"pearson_cond"}: Pearson conditional residuals
  \item \code{"pearson_mar"}: Pearson marginal residuals
  \item \code{"raw_cond"}: raw conditional residuals (default)
  \item \code{"raw_mar"}: raw marginal residuals
  \item \code{"std_cond"}: studentized conditional residuals
  \item \code{"std_mar"}: studentized marginal residuals
}
See the vignette for details on how residual types are computed.
}
\examples{
# fits a linear mixed effects model
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), data = sleepstudy)

# computes the default residuals (raw conditional)
redres(fm1)

# changes the residual type to studentized marginal
redres(fm1, type = "std_mar")
}
