% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_paprika.R
\docType{data}
\name{paprika}
\alias{paprika}
\title{Paprika Dataset}
\format{Tibble}
\usage{
paprika
}
\description{
Data from a study on the growth rate of paprika plants grown under small-scale farming
conditions in southern Africa. The study was conducted in 2007 and 2008 in Malawi. The
researcher compared 4 fertilizer treatments and 6 plant varieties.
}
\details{
((1) control, (2) inorganic fertilizer,
(3) Gliricidia biomass, and (4) Gliricidia biomass plus half the recommended dose of
fertilizer for paprika) and 6 varieties of plant ((1) Papri-King, (2) Papri-Queen,
(3) Papri-Excel, (4) Papri-Supreme, (5) Papri-Ace, and (6) PX1140 4601.

The
experimental design of the study was a split-plot with 5 replicates.

This dataset contains observations from week 4 of the
study in 2008.

The variables in the dataset are as follows.
\tabular{ll}{
 \code{heartrate} \tab Heart rate of the penguin during the dive (beats per minute) \cr
 \code{depth} \tab Depth of the dive (meters) \cr
 \code{duration} \tab Duration of the dive (minutes) \cr
 \code{bird} \tab The id number associated with a penguin}
}
\keyword{datasets}
