% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genres.R
\name{plot_genres}
\alias{plot_genres}
\title{Generalized residual quantile plot}
\usage{
plot_genres(model)
}
\arguments{
\item{model}{Model fit using \code{lmer}.}
}
\description{
Use to compare the generalized residuals to the uniform quantiles. Obvious departures indicate
a lack of fit.
}
\details{
First section what is a generalized residual. Reiterate diagnostics.
Second section assumptions of lmer model
}
\examples{
# Fit a linear mixed effect model using lme4 where model has a
# random intercept for Days and random slope for Subject*Days.
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
# Check that all model assumptions are reasonable using generalized residuals:
plot_genres(fm1)
}
