% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genres.R
\name{plot_genres}
\alias{plot_genres}
\title{Generalized residual quantile plot}
\usage{
plot_genres(model)
}
\arguments{
\item{model}{Model fit using \code{lmer}.}
}
\description{
Use to compare the generalized residuals to the uniform quantiles. Obvious departures indicate
a lack of fit.
}
\details{
First section what is a generalized residual. Reiterate diagnostics.
Second section assumptions of lmer model
}
