% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raneff.R
\name{plot_raneff}
\alias{plot_raneff}
\title{Normal quantile plots for random effects}
\usage{
plot_raneff(model)
}
\arguments{
\item{model}{Model fit using \code{lmer}.}
}
\description{
Use to compare the random effects to normal quantiles. How to use for diagnosing...
}
\details{
Why random effects should be normally distributed and diagnosing qq-plot departures.
}
