% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_redres.R
\name{plot_redres}
\alias{plot_redres}
\title{Diagnostic residual plot for linear mixed models}
\usage{
plot_redres(model, type = "raw_cond", xvar = NULL)
}
\arguments{
\item{model}{Model fit using \code{lmer} from \code{lme4}.}

\item{type}{String identifying type of residual. Default is "raw_cond".
See \code{\link{redres}} for details of available types.}

\item{xvar}{String indicates the variable to be plotted at the x-axis. By default,
the fitted values are plotted on the x-axis. Any variable used in the
lmer model can be specified.}
}
\value{
A residual plot in the form of a \code{ggplot2} object.
}
\description{
Creates a plot of residuals versus fitted values or model variable. This plot can be used to assess whether the assumptions of constant
variance and linear form assumptions are adequate.
}
\examples{
# fits a linear mixed effects model
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)

# plots raw conditional residuals by fitted values.
plot_redres(fm1)

# plots raw conditional residuals by selected variables `Days`.
plot_redres(fm1, xvar = "Days")

# plots standardized conditional residuals by fitted values.
plot_redres(fm1, type = "std_cond")

# edits theme of ggplot2 object
plot_redres(fm1, type = "pearson_mar") + theme_grey()
}
