% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resid.R
\name{plot_redres}
\alias{plot_redres}
\title{Diagnostic residual plot for linear mixed models}
\usage{
plot_redres(model, type = "raw_cond")
}
\arguments{
\item{model}{Model fit using \code{lmer}.}

\item{type}{String identifying type of residual. Default is "raw_cond".
See \code{\link{redres}} for details of available types.
Note that the type of \code{genres} is not available for this plot
since it is not meaningful for generalized residuals.}
}
\value{
A plot of residuals versus fitted values.
}
\description{
Plot of specified residual type versus fitted values for a linear mixed effects model
fitted using \code{lmer}. Use plot to assess whether the assumptions of constant
variance and linear form are adequate.
}
\examples{
# Fit a linear mixed effect model with a default (raw conditional) residual type.
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)

# Plot raw conditional residuals by fitted values.
plot_redres(fm1)

# Plot standardized conditional residuals by fitted values.
plot_redres(fm1, type = "std_cond")
}
