% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resqq.R
\name{plot_resqq}
\alias{plot_resqq}
\title{Normal quantile plot for residuals.}
\usage{
plot_resqq(model)
}
\arguments{
\item{model}{Model fit using \code{lmer}.}
}
\value{
A generalized residual quantile plot.
}
\description{
Creates a normal quantile plot of the raw conditional residuals. By the assumptions
of a model fit using \code{lmer} these residuals are expected to be normally distributed.
}
\details{
Confidence bands are constructed from pointwise normal confidence intervals.
}
\examples{
# fits a linear mixed effect model using lme4 where model has a
# random intercept for Days and random slope for Subject*Days
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)

# check that error term is normally distributed
plot_resqq(fm1)
}
