% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{redres}
\alias{redres}
\title{Compute the Residuals}
\usage{
redres(model, type = "raw_cond")
}
\arguments{
\item{model}{the model we fit from lmer function in lme4 package.}

\item{type}{type of residuals. we have generalized, standardized, conditional and pearson residuals.}
}
\value{
return residuals according to different types.
}
\description{
function to compute residuals. We we have four types methods to caculate residuals,
which are pearson residuals, generalized residuals, standardized residuals and
conditional residuals.
}
\examples{
#fit the model
library(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), data = sleepstudy)
#compute the residuals
result <- redres(fm1, type = "raw_cond")
}
